prepare for new release JX-1.7.0-5
I am calling the next release JX-1.7.0-5 so that we actually know about
which release we are talking. The pre-release version is 1.7.0-4 based.

The difference between all the previous versions and this release is
that this is based on the open source release which are available
through developerWorks:

http://www.ibm.com/developerworks/power/pa-slof/

To build an actual release there are also some binary only files
required which can be downloaded at:

http://www.ibm.com/developerworks/power/pa-slof-js20/