Improved make.rules files
The make.rules of net-snk now uses the rules of the main make.rules file, too,
so that it is not necessary anymore to configure everything twice.
Also replaced the non-intuitive NEW_BUILD variable with the simply "V" flag
variable, so that you can now type "make V=x" with x being either 0, 1 or 2
for controlling the verbosity level of the build process.
Finally I also replaced the "echo -e" constructs with printfs since "echo -e"
is not POSIX compliant and and thus does not work other shells than bash.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>