Fixed paflof Makefile
Replaced the \t in the sed rules with real tabs since this is more portable
(\t only works with the GNU version of sed, not with the BSD version).
Also replaced the pushd and popd commands since they are not POSIX compliant,
they only work with bash but e.g. not with dash on Debian/Ubuntu systems.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>