First fixes for the OHCI code on board-qemu.
The emulated OHCI controller on board-qemu behaves slightly different to the
controller that was used on JS2x. We've got to translate the PCI base address
to get the correct virtual base address of the HC. QEMU also does not like
mis-aligned accesses to registers, and after powering off the USB ports, we
must make sure to issue a reset to be able to detect devices again.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>