Fix for 8-byte wide local-mac-address properties
The /vdevice/l-lan device tree node from qemu uses a 8-byte wide
local-mac-address property instead of a normal 6-byte wide one.
According to a comments in the qemu code, this is necessary for backward
compatibility with old Linux kernels, so the problem is not going to be fixed
on the qemu side and we've got to take care about this in SLOF instead.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>