Improved the way to detect some execution tokens for the debugger.
The debugger used an ugly hack to determine the execution tokens for
2CONSTANT, INSTANCE VALUE and INSTANCE VARIABLE. This hack used an faked
device node, which was quite error prone, e.g. when extending the node
structure in node.fs. Now the XTs are determined in a much more elegant way
at the places where the corresponding Forth words are defined.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>