Fixed PACK command.
According to IEEE 1275, the "pack" command converts a Forth string into a
packed counted string, i.e. it has to store the length of the string in the
destination memory buffer, too, so that the operation can be reverted with the
"count" command. Our current implementation did not save the length yet - this
has been fixed now.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>