Added net-snk client-interface module
So far the net-snk could only use networking cards that had corresponding
drivers in the ROM filesystem. NICs that come with their of FCODE drivers were
not supported yet. For these NICs I added now a pseudo-module to the net-snk
that is able to use the corresponding device tree nodes via the "read" and
"write" functions of the client interface.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>