Fixed "dev" command to only parse one word
The "dev" command parsed the whole remaining input line. That way it was not
possible to do something like "dev / ls" at the prompt. Now it parses only
the next available word in the input so that the rest of the input line can
be interpreted by the Forth engine instead.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>