Two minor improvements for the device tree node code
The "open-node" command always tried to execute the "open" method of the node
that should be opened and used CATCH to determine whether the method is
available or not (and assumed success if it is not available). That works fine
as long as the "open" method does not ABORT due to any reason - in that case
the "open" should be considered as failure instead. So the code is now using
"find-method" instead to determine whether the "open" method is available or
not.
Second fix is about "my-unit". According to IEEE1275, this value should be
initialized to the firt component of the "reg" property in case it has not
been specified by other means.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>