Fixed yet another bug in open-node
When the "open" method of the target node failed, the open-node function did
not correctly restore the previous value of "my-self" (it was using my-parent
of the new node to restore the previous value, but that did not work when the
new node could not be opened). Now the old value of "my-self" is saved on the
return stack instead and then always correctly restored.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>