usb-ohci: Avoid instance specific allocations
The USB stack did instance specific allocations for the
various buffers, but would then assign them to package
specific lists, and would generally get badly confused
as soon as multiple instances of the driver got opened.

This typically occurs when using a USB keyboard and booting
from a USB disk.

This makes everything node variables instead, with tracking
of how many instances are open, with one allocation pass
on the first open and one free pass on the last close.

The result is a lot more reliable, I can boot with USB keyboard
and a USB disk fine without hanging qemu.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>