Fixes to ANSI escape sequence handling
We have a couple of issues making grub2 essentially unusable when
using a framebuffer console:

 - When we encounter a ";" in a CSI sequence, we still stop the
parsing, so all the positioning sequences are broken

 - We made a pretty poor choice of when to use invert video,
basically switching to invert whenever a "m" sequence had any
non-zero value. Instead we remember the foreground and background
colors set by the client and use invert when the foreground value
is lower (or equal) to the background value.

There is still a -lot- of improvements to do here (including proper
support for colors) but at least this makes grub2 from fedora 17
usable.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>