Workaround for SLES11 installer bug
This is a hack to work around some stupid behaviour in the SLES11
installer.  It expects a node named simply "disk" under the VIO SCSI
device node, rather than individual disk@XXX devices.  pHyp doesn't
enumerate the disks, and so provides such a node with the disk only probed
when the device is actually opened.  SLOF does enumerate the disks and so
provides the nodes with specific unit addresses instead, breaking the
installer.

The installer doesn't actually use anything in the "disk" node however, it
just expects it to be there, so this hack provides a dummy node to make
the installer happy.

Signed-off-by: Ben Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>