SLOF: vio-vscsi: fixed bug with reported luns
The vscsi-report-luns scans up to 8 devices and returns
a number of found devices and an array of pointers, one per device.
Each pointer points to a null-terminated list of LUNs reported by
a device in a response to REPORT_LUNS SCSI command.

The vscsi-find-disks function uses the result returned vscsi-report-luns
to walk through the lists and probe all the devices. However the loop
through a null terminated list was interrupted on the first iteration.
The patch fixes this bug.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Signed-off-by: Nikunj A Dadhania <nikunj@linux.vnet.ibm.com>