vio-vscsi: Switch to using a wildcard "disk" node and make scsi-disk generic
This switches vio-vscsi to using a wildcard "disk" node instead of
creating nodes for all children, which is more in line with what
other OFW implementations do and fixes issues with some clients
(such as the SLES11SP2 installer) in a better way than the previous
hack of creating a non-functional disk device.

Additionally, the vio-vscsi-device.fs now becomes a generic scsi-disk.fs
which should generally be usable underneath any scsi-style transport
that provides a handful of methods.

Still to do to be come really generic:

Simplify interface to the parent by implementing an "execute-command"
akin to what is described in the OF SCSI binding and move the actual
construction of most SCSI commands into scsi-disk.fs itself.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>