Add hack to client interface finddevice of "/memory"
grub2 is bogus, it looks for "/memory" to check for available
memory. On a system with multiple memory nodes that might return
anything and not necessarily the RMA node which is the only one
usable in real mode.

This adds a hack to the client interface code which transforms
any request for "/memory" to explicitly look for "/memory@0"
which returns the RMA node.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>