usb: Use proper accessors for MMIO and separate in-memory endian swap
We cannot just access MMIO like that. On KVM for example, this has
to be hypercalls. On js2x, we need to use special cache-inhibited
loads and stores.

We have accessors in cache.h, we just need to use them. However that
means that read/write_reg() are no longer suitable for in-memory
byteswaps. We need to use the accessors in byteorder.h for that

While at it, we also fix the code to use mb() instead of barrier()
as a full memory barrier is needed to actually order things.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>