Use root.fs on qemu as well
We had lots of duplication of code between root.fs and fdt.fs

The reason we didn't use root.fs in the first place is that it tried
to create the root node and the chosen node which we already have via
the fdt.

Instead, modify root.fs to check for their presence and only "extend"
them in that case.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>