Change representation of string environment variable
Currently strings are stored in the dictionary using "string," which
stores the length as a byte. This brings in the limitation of only 255
byte string. The issue got exposed when using "boot-device" which was
more than 256 bytes, and the string length was equal to (length & 0xFF)

With this patch, we are storing the length in dictionary in a cell and
copying the string after that.

Signed-off-by: Nikunj A Dadhania <nikunj@linux.vnet.ibm.com>