cas: increase hcall buffer size to accomodate 256 cpus
It is seen that guests booting in compatibility mode (like Power7)
will issue a client-architecture-service call which will result in
QEMU rebuilding the device tree. The size of the device tree to be
built, which is passed from SLOF is currently 4 kB. This is
insufficient to contain more than 48 CPU nodes. Increase this size in
SLOF so that 255 CPUs (max limit in QEMU) can be supported.

Signed-off-by: Nikunj A Dadhania <nikunj@linux.vnet.ibm.com>