pci-bridge: Fix ranges when no device beyond the bridge
Before starting pci-bridge probing the ranges property is set to the
max limit, and then pci bridge is probed. Once the probe is over we
would update the ranges property.

In case when there is no device beyond the bridge, ranges will be null
and the code moves ahead. The actual problem here is that before
probing the bridge the ranges property was set to max limit. That
stays and cause overlapping ranges for sibling devices.

So empty the ranges property for this bridge when there is no device present.

Signed-off-by: Nikunj A Dadhania <nikunj@linux.vnet.ibm.com>