Fix term-io-key to also work when stdin has not been set yet
term-io-key simply did nothing when the stdin had not been set yet.
Especially, it even did not put an output item on the stack, which led
to very ugly crashes in case we dropped to the Forth prompt before
the console had been set up. Fix it by running the behaviour of the
previous key handler instead.

Signed-off-by: Thomas Huth <thuth@linux.vnet.ibm.com>
Signed-off-by: Nikunj A Dadhania <nikunj@linux.vnet.ibm.com>