pci: scan only type 0 and type 1
In certain cases when a PCI device is passthru and is frozen before it
reaches SLOF.  All PCI read result reading FF. This exposed a bug in
the pci scan code which would wrongly take it as a pci brigde and try
to scan devices below that and that would enter an infinite loop.

Explicitly scan for type 0/1 and rest are not supported.

Signed-off-by: Nikunj A Dadhania <nikunj@linux.vnet.ibm.com>